requirements.txt
kafka-python
fastapi
uvicorn
pydantic-settings
